Make allocator thread-safe
Separate Batch definition from Batcher
